LayoutTests/storage/indexeddb/modern/opendatabase-request-event-expected.txt

ALERT: Caught exception TypeError: Type error
ALERT: Caught exception TypeError: Type error
ALERT: [object IDBOpenDBRequest]
ALERT: upgradeneeded: old version - 0 new version - 1
ALERT: Done
This test calls open on window.indexedDB in various ways, verifying that IDBOpenDBRequest objects are returned when expected and exceptions are thrown when expected.
